Adding flexibility to mock ResultSets
