+++ java/adapter/jdbc/src/test/java/org/apache/arrow/adapter/jdbc/h2/JdbcToArrowTest.java
@@ -43,30 +43,12 @@ import static org.apache.arrow.adapter.jdbc.JdbcToArrowTestHelper.getLongValues;
 import static org.junit.Assert.*;
 
 import java.io.IOException;
-import java.io.InputStream;
-import java.io.Reader;
-import java.math.BigDecimal;
-import java.net.URL;
-import java.sql.Array;
-import java.sql.Blob;
-import java.sql.Clob;
-import java.sql.Date;
-import java.sql.NClob;
-import java.sql.Ref;
 import java.sql.ResultSet;
 import java.sql.ResultSetMetaData;
-import java.sql.RowId;
 import java.sql.SQLException;
-import java.sql.SQLWarning;
-import java.sql.SQLXML;
-import java.sql.Statement;
-import java.sql.Time;
-import java.sql.Timestamp;
-import java.sql.Types;
 import java.util.Arrays;
 import java.util.Calendar;
 import java.util.Collection;
-import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
@@ -77,6 +59,7 @@ import org.apache.arrow.adapter.jdbc.JdbcToArrowConfig;
 import org.apache.arrow.adapter.jdbc.JdbcToArrowConfigBuilder;
 import org.apache.arrow.adapter.jdbc.JdbcToArrowTestHelper;
 import org.apache.arrow.adapter.jdbc.JdbcToArrowUtils;
+import org.apache.arrow.adapter.jdbc.ResultSetUtility;
 import org.apache.arrow.adapter.jdbc.Table;
 import org.apache.arrow.memory.BufferAllocator;
 import org.apache.arrow.memory.RootAllocator;
@@ -234,7 +217,7 @@ public class JdbcToArrowTest extends AbstractJdbcToArrowTest {
     BufferAllocator allocator = new RootAllocator(Integer.MAX_VALUE);
     int x = 0;
     final int targetRows = 600000;
-    ResultSet rs = new FakeResultSet(targetRows);
+    ResultSet rs = ResultSetUtility.generateBasicResultSet(targetRows);
     JdbcToArrowConfig config = new JdbcToArrowConfigBuilder(
         allocator, JdbcToArrowUtils.getUtcCalendar(), /* include metadata */ false)
         .setReuseVectorSchemaRoot(reuseVectorSchemaRoot).build();
@@ -259,7 +242,7 @@ public class JdbcToArrowTest extends AbstractJdbcToArrowTest {
     BufferAllocator allocator = new RootAllocator(Integer.MAX_VALUE);
     int x = 0;
     final int targetRows = 0;
-    ResultSet rs = new FakeResultSet(targetRows);
+    ResultSet rs = ResultSetUtility.generateEmptyResultSet();
     JdbcToArrowConfig config = new JdbcToArrowConfigBuilder(
             allocator, JdbcToArrowUtils.getUtcCalendar(), /* include metadata */ false)
             .setReuseVectorSchemaRoot(reuseVectorSchemaRoot).build();
@@ -268,1102 +251,4 @@ public class JdbcToArrowTest extends AbstractJdbcToArrowTest {
     assertFalse("Iterator on zero row ResultSet should not haveNext()", iter.hasNext());
   }
 
-  private class FakeResultSet implements ResultSet {
-
-    public int numRows;
-    public boolean empty;
-
-    FakeResultSet(int numRows) {
-      this.numRows = numRows;
-      this.empty = numRows <= 0;
-    }
-
-    @Override
-    public boolean next() throws SQLException {
-      numRows--;
-      return numRows >= 0;
-    }
-
-    @Override
-    public void close() throws SQLException {
-
-    }
-
-    @Override
-    public boolean wasNull() throws SQLException {
-      return false;
-    }
-
-    @Override
-    public String getString(int columnIndex) throws SQLException {
-      return "test123test123" + numRows;
-    }
-
-    @Override
-    public boolean getBoolean(int columnIndex) throws SQLException {
-      return false;
-    }
-
-    @Override
-    public byte getByte(int columnIndex) throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public short getShort(int columnIndex) throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public int getInt(int columnIndex) throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public long getLong(int columnIndex) throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public float getFloat(int columnIndex) throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public double getDouble(int columnIndex) throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public BigDecimal getBigDecimal(int columnIndex, int scale) throws SQLException {
-      return new BigDecimal(5);
-    }
-
-    @Override
-    public byte[] getBytes(int columnIndex) throws SQLException {
-      return new byte[0];
-    }
-
-    @Override
-    public Date getDate(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Time getTime(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Timestamp getTimestamp(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public InputStream getAsciiStream(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public InputStream getUnicodeStream(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public InputStream getBinaryStream(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public String getString(String columnLabel) throws SQLException {
-      throw new UnsupportedOperationException("get column by label not supported");
-    }
-
-    @Override
-    public boolean getBoolean(String columnLabel) throws SQLException {
-      return false;
-    }
-
-    @Override
-    public byte getByte(String columnLabel) throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public short getShort(String columnLabel) throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public int getInt(String columnLabel) throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public long getLong(String columnLabel) throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public float getFloat(String columnLabel) throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public double getDouble(String columnLabel) throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public BigDecimal getBigDecimal(String columnLabel, int scale) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public byte[] getBytes(String columnLabel) throws SQLException {
-      return new byte[0];
-    }
-
-    @Override
-    public Date getDate(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Time getTime(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Timestamp getTimestamp(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public InputStream getAsciiStream(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public InputStream getUnicodeStream(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public InputStream getBinaryStream(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public SQLWarning getWarnings() throws SQLException {
-      return null;
-    }
-
-    @Override
-    public void clearWarnings() throws SQLException {
-
-    }
-
-    @Override
-    public String getCursorName() throws SQLException {
-      return null;
-    }
-
-    @Override
-    public ResultSetMetaData getMetaData() throws SQLException {
-      return new ResultSetMetaData() {
-        @Override
-        public int getColumnCount() throws SQLException {
-          return 5;
-        }
-
-        @Override
-        public boolean isAutoIncrement(int column) throws SQLException {
-          return false;
-        }
-
-        @Override
-        public boolean isCaseSensitive(int column) throws SQLException {
-          return false;
-        }
-
-        @Override
-        public boolean isSearchable(int column) throws SQLException {
-          return false;
-        }
-
-        @Override
-        public boolean isCurrency(int column) throws SQLException {
-          return false;
-        }
-
-        @Override
-        public int isNullable(int column) throws SQLException {
-          return 0;
-        }
-
-        @Override
-        public boolean isSigned(int column) throws SQLException {
-          return false;
-        }
-
-        @Override
-        public int getColumnDisplaySize(int column) throws SQLException {
-          return 0;
-        }
-
-        @Override
-        public String getColumnLabel(int column) throws SQLException {
-          return getColumnName(column);
-        }
-
-        @Override
-        public String getColumnName(int column) throws SQLException {
-          return "col_" + column;
-        }
-
-        @Override
-        public String getSchemaName(int column) throws SQLException {
-          return null;
-        }
-
-        @Override
-        public int getPrecision(int column) throws SQLException {
-          return 0;
-        }
-
-        @Override
-        public int getScale(int column) throws SQLException {
-          return 0;
-        }
-
-        @Override
-        public String getTableName(int column) throws SQLException {
-          return null;
-        }
-
-        @Override
-        public String getCatalogName(int column) throws SQLException {
-          return null;
-        }
-
-        @Override
-        public int getColumnType(int column) throws SQLException {
-          switch (column) {
-            case 1:
-              return Types.VARCHAR;
-            case 2:
-              return Types.INTEGER;
-            case 3:
-              return Types.BIGINT;
-            case 4:
-              return Types.VARCHAR;
-            case 5:
-              return Types.VARCHAR;
-            default:
-              throw new IllegalArgumentException("not supported");
-          }
-
-        }
-
-        @Override
-        public String getColumnTypeName(int column) throws SQLException {
-          return null;
-        }
-
-        @Override
-        public boolean isReadOnly(int column) throws SQLException {
-          return false;
-        }
-
-        @Override
-        public boolean isWritable(int column) throws SQLException {
-          return false;
-        }
-
-        @Override
-        public boolean isDefinitelyWritable(int column) throws SQLException {
-          return false;
-        }
-
-        @Override
-        public String getColumnClassName(int column) throws SQLException {
-          return null;
-        }
-
-        @Override
-        public <T> T unwrap(Class<T> iface) throws SQLException {
-          return null;
-        }
-
-        @Override
-        public boolean isWrapperFor(Class<?> iface) throws SQLException {
-          return false;
-        }
-      };
-    }
-
-    @Override
-    public Object getObject(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Object getObject(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public int findColumn(String columnLabel) throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public Reader getCharacterStream(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Reader getCharacterStream(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public BigDecimal getBigDecimal(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public BigDecimal getBigDecimal(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public boolean isBeforeFirst() throws SQLException {
-      return false;
-    }
-
-    @Override
-    public boolean isAfterLast() throws SQLException {
-      if (empty) {
-        return false;
-      }
-      return numRows < 0;
-    }
-
-    @Override
-    public boolean isFirst() throws SQLException {
-      return false;
-    }
-
-    @Override
-    public boolean isLast() throws SQLException {
-      return numRows <= 0;
-    }
-
-    @Override
-    public void beforeFirst() throws SQLException {
-
-    }
-
-    @Override
-    public void afterLast() throws SQLException {
-
-    }
-
-    @Override
-    public boolean first() throws SQLException {
-      return false;
-    }
-
-    @Override
-    public boolean last() throws SQLException {
-      return false;
-    }
-
-    @Override
-    public int getRow() throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public boolean absolute(int row) throws SQLException {
-      return false;
-    }
-
-    @Override
-    public boolean relative(int rows) throws SQLException {
-      return false;
-    }
-
-    @Override
-    public boolean previous() throws SQLException {
-      return false;
-    }
-
-    @Override
-    public void setFetchDirection(int direction) throws SQLException {
-
-    }
-
-    @Override
-    public int getFetchDirection() throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public void setFetchSize(int rows) throws SQLException {
-
-    }
-
-    @Override
-    public int getFetchSize() throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public int getType() throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public int getConcurrency() throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public boolean rowUpdated() throws SQLException {
-      return false;
-    }
-
-    @Override
-    public boolean rowInserted() throws SQLException {
-      return false;
-    }
-
-    @Override
-    public boolean rowDeleted() throws SQLException {
-      return false;
-    }
-
-    @Override
-    public void updateNull(int columnIndex) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBoolean(int columnIndex, boolean x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateByte(int columnIndex, byte x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateShort(int columnIndex, short x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateInt(int columnIndex, int x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateLong(int columnIndex, long x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateFloat(int columnIndex, float x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateDouble(int columnIndex, double x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBigDecimal(int columnIndex, BigDecimal x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateString(int columnIndex, String x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBytes(int columnIndex, byte[] x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateDate(int columnIndex, Date x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateTime(int columnIndex, Time x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateTimestamp(int columnIndex, Timestamp x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateAsciiStream(int columnIndex, InputStream x, int length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBinaryStream(int columnIndex, InputStream x, int length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateCharacterStream(int columnIndex, Reader x, int length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateObject(int columnIndex, Object x, int scaleOrLength) throws SQLException {
-
-    }
-
-    @Override
-    public void updateObject(int columnIndex, Object x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateNull(String columnLabel) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBoolean(String columnLabel, boolean x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateByte(String columnLabel, byte x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateShort(String columnLabel, short x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateInt(String columnLabel, int x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateLong(String columnLabel, long x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateFloat(String columnLabel, float x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateDouble(String columnLabel, double x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBigDecimal(String columnLabel, BigDecimal x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateString(String columnLabel, String x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBytes(String columnLabel, byte[] x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateDate(String columnLabel, Date x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateTime(String columnLabel, Time x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateTimestamp(String columnLabel, Timestamp x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateAsciiStream(String columnLabel, InputStream x, int length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBinaryStream(String columnLabel, InputStream x, int length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateCharacterStream(String columnLabel, Reader reader, int length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateObject(String columnLabel, Object x, int scaleOrLength) throws SQLException {
-
-    }
-
-    @Override
-    public void updateObject(String columnLabel, Object x) throws SQLException {
-
-    }
-
-    @Override
-    public void insertRow() throws SQLException {
-
-    }
-
-    @Override
-    public void updateRow() throws SQLException {
-
-    }
-
-    @Override
-    public void deleteRow() throws SQLException {
-
-    }
-
-    @Override
-    public void refreshRow() throws SQLException {
-
-    }
-
-    @Override
-    public void cancelRowUpdates() throws SQLException {
-
-    }
-
-    @Override
-    public void moveToInsertRow() throws SQLException {
-
-    }
-
-    @Override
-    public void moveToCurrentRow() throws SQLException {
-
-    }
-
-    @Override
-    public Statement getStatement() throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Object getObject(int columnIndex, Map<String, Class<?>> map) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Ref getRef(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Blob getBlob(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Clob getClob(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Array getArray(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Object getObject(String columnLabel, Map<String, Class<?>> map) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Ref getRef(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Blob getBlob(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Clob getClob(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Array getArray(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Date getDate(int columnIndex, Calendar cal) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Date getDate(String columnLabel, Calendar cal) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Time getTime(int columnIndex, Calendar cal) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Time getTime(String columnLabel, Calendar cal) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Timestamp getTimestamp(int columnIndex, Calendar cal) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Timestamp getTimestamp(String columnLabel, Calendar cal) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public URL getURL(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public URL getURL(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public void updateRef(int columnIndex, Ref x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateRef(String columnLabel, Ref x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBlob(int columnIndex, Blob x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBlob(String columnLabel, Blob x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateClob(int columnIndex, Clob x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateClob(String columnLabel, Clob x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateArray(int columnIndex, Array x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateArray(String columnLabel, Array x) throws SQLException {
-
-    }
-
-    @Override
-    public RowId getRowId(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public RowId getRowId(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public void updateRowId(int columnIndex, RowId x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateRowId(String columnLabel, RowId x) throws SQLException {
-
-    }
-
-    @Override
-    public int getHoldability() throws SQLException {
-      return 0;
-    }
-
-    @Override
-    public boolean isClosed() throws SQLException {
-      return false;
-    }
-
-    @Override
-    public void updateNString(int columnIndex, String nString) throws SQLException {
-
-    }
-
-    @Override
-    public void updateNString(String columnLabel, String nString) throws SQLException {
-
-    }
-
-    @Override
-    public void updateNClob(int columnIndex, NClob nClob) throws SQLException {
-
-    }
-
-    @Override
-    public void updateNClob(String columnLabel, NClob nClob) throws SQLException {
-
-    }
-
-    @Override
-    public NClob getNClob(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public NClob getNClob(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public SQLXML getSQLXML(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public SQLXML getSQLXML(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public void updateSQLXML(int columnIndex, SQLXML xmlObject) throws SQLException {
-
-    }
-
-    @Override
-    public void updateSQLXML(String columnLabel, SQLXML xmlObject) throws SQLException {
-
-    }
-
-    @Override
-    public String getNString(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public String getNString(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Reader getNCharacterStream(int columnIndex) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public Reader getNCharacterStream(String columnLabel) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public void updateNCharacterStream(int columnIndex, Reader x, long length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateNCharacterStream(String columnLabel, Reader reader, long length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateAsciiStream(int columnIndex, InputStream x, long length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBinaryStream(int columnIndex, InputStream x, long length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateCharacterStream(int columnIndex, Reader x, long length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateAsciiStream(String columnLabel, InputStream x, long length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBinaryStream(String columnLabel, InputStream x, long length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateCharacterStream(String columnLabel, Reader reader, long length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBlob(int columnIndex, InputStream inputStream, long length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBlob(String columnLabel, InputStream inputStream, long length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateClob(int columnIndex, Reader reader, long length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateClob(String columnLabel, Reader reader, long length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateNClob(int columnIndex, Reader reader, long length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateNClob(String columnLabel, Reader reader, long length) throws SQLException {
-
-    }
-
-    @Override
-    public void updateNCharacterStream(int columnIndex, Reader x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateNCharacterStream(String columnLabel, Reader reader) throws SQLException {
-
-    }
-
-    @Override
-    public void updateAsciiStream(int columnIndex, InputStream x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBinaryStream(int columnIndex, InputStream x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateCharacterStream(int columnIndex, Reader x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateAsciiStream(String columnLabel, InputStream x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBinaryStream(String columnLabel, InputStream x) throws SQLException {
-
-    }
-
-    @Override
-    public void updateCharacterStream(String columnLabel, Reader reader) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBlob(int columnIndex, InputStream inputStream) throws SQLException {
-
-    }
-
-    @Override
-    public void updateBlob(String columnLabel, InputStream inputStream) throws SQLException {
-
-    }
-
-    @Override
-    public void updateClob(int columnIndex, Reader reader) throws SQLException {
-
-    }
-
-    @Override
-    public void updateClob(String columnLabel, Reader reader) throws SQLException {
-
-    }
-
-    @Override
-    public void updateNClob(int columnIndex, Reader reader) throws SQLException {
-
-    }
-
-    @Override
-    public void updateNClob(String columnLabel, Reader reader) throws SQLException {
-
-    }
-
-    @Override
-    public <T> T getObject(int columnIndex, Class<T> type) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public <T> T getObject(String columnLabel, Class<T> type) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public <T> T unwrap(Class<T> iface) throws SQLException {
-      return null;
-    }
-
-    @Override
-    public boolean isWrapperFor(Class<?> iface) throws SQLException {
-      return false;
-    }
-  }
 }
